Kernel/Net: Move NE2000 network adapter code to a separate directory
