+++ Kernel/Net/NE2000/NetworkAdapter.h
